socket_server/pyproject.toml
[project]
name = "garden-socket-server"
version = "0.1.0"
description = "Socket server for garden project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt",
]

[tool.uv]
managed = true
